fix some extern bug
