+++ cinn/runtime/custom_function.h
@@ -62,7 +62,5 @@ class AssertTrueMsgTool {
 
 void cinn_assert_true(void* v_args, int num_args, int msg, bool only_warning, void* stream, const Target& target);
 
-void cinn_call_cholesky_host(void* v_args, int num_args, int batch_size, int m, bool upper);
-
 }  // namespace runtime
 }  // namespace cinn
