+++ cinn/runtime/custom_function_test.cc
@@ -24,6 +24,10 @@
 #include "cinn/runtime/cuda/cuda_util.h"
 #endif
 
+#ifdef CINN_WITH_MKL_CBLAS
+#include "cinn/runtime/cpu/cblas.h"
+#endif
+
 #include "cinn/runtime/cinn_runtime.h"
 #include "cinn/runtime/custom_function.h"
